requirements.txt
fastapi
uvicorn
python-multipart
spotipy
pydub
numpy==1.25.2
python-dotenv
requests
tensorflow==2.15.0
tensorflow_hub==0.16.1
joblib==1.3.2
scipy==1.11.4
scikit-learn==1.2.2